+++ src/app/View/Drawer/ManageTags/context/useDeleteTag.ts
@@ -1,0 +1,5 @@
+import { useDeleteTag as useDeleteTagAction } from "../../../../ActionCreator/tags/useDeleteTag";
+
+export const useDeleteTag = () => {
+  return useDeleteTagAction();
+}
