+++ src/app/View/Modal/Modal.tsx
@@ -15,12 +15,12 @@ import FooterItemEditor from "../Drawer/FooterItemEditor/view/FooterItemEditor";
 import { FeedbackForm } from "../Drawer/FeedbackForm/view/FeedbackForm";
 import { ManageTheme } from "../Drawer/ManageTheme/view/ManageTheme";
 import { SelectTags } from "../Footer/SelectTags";
-import { ManageTags } from "../Drawer/ManageTags/ManageTags";
+import { ManageTags } from "../Drawer/ManageTags/view/ManageTags";
 import { SettingUserInfo } from "../Drawer/Account/view/ManageUserInfo";
 import { DeleteAccountForm } from "../Drawer/Account/view/DeleteAccountForm";
 import { useModalSize, medium } from "./useModalSize";
 import { StyledDialog } from "./StyledDialog";
-import { ManageInstagramAccounts } from "../Drawer/ManageInstagramAccounts/ManageInstagmaAccounts";
+import { ManageInstagramAccounts } from "../Drawer/ManageInstagramAccounts/view/ManageInstagmaAccounts";
 import { SelectInstagramAccounts } from "../Footer/SelectInstagramAccounts";
 import { TSetModal, T_selected_theme } from "../../Store/Types";
 import { InstagramMediaModal } from "../Main/InstagramMediaModal";
